Infinite loop in SourceBufferParserWebM::SegmentReader::Read https://bugs.webkit.org/show_bug.cgi?id=232740 <rdar://83573873>
Reviewed by Eric Carlson.

Source/WebCore:

No test because the reproduction steps are still unknown.

Customer reports indicate an infinite loop condition on a background queue inside
SourceBufferParserWebM::SegmentReader::Read(), and those reports indicate that the
MTPluginByteSourceRead() call is failing and returning an error. When an error occurs, the
SourceBufferParser::Segment::read() method drops the error and simply reports that zero
bytes were read. And consequently in the SegmentReader::Read() method, that method will
loop forever attempting to satisfy the request for a certain number of bytes to be read.

The cause of the MTPluginByteSourceRead() call failing is not well known at this point.
It could be the case of a faulty server reporting an incorrect content-length, or it could
be a result of a WebContent process crash. Regardless, that error should be propagated
up the call stack so that the caller can correctly break out of its while loop.

Change the return type the Segment::read() method from `size_t` to an Expected object
containing either a `size_t` or a ReaderError. Check the return value from within
SegmentReader::Read() and bail if an error was returned. ReadInto() will need to be
modified similarly.

* platform/graphics/cocoa/SourceBufferParser.cpp:
(WebCore::SourceBufferParser::Segment::read const):
(WebCore::SourceBufferParser::Segment::takeSharedBuffer):
* platform/graphics/cocoa/SourceBufferParser.h:
* platform/graphics/cocoa/SourceBufferParserWebM.cpp:
(WebCore::segmentReadErrorToWebmStatus):
* platform/graphics/cocoa/SourceBufferParserWebM.h:

Source/WebCore/PAL:

* pal/spi/cocoa/MediaToolboxSPI.h:



Canonical link: https://commits.webkit.org/243959@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285402 268f45cc-cd09-0410-ab3c-d52691b4dbfc