[GPU Process] Remove non-platform code form the FEImage class https://bugs.webkit.org/show_bug.cgi?id=232369 rdar://84966765
Reviewed by Simon Fraser.

Source/WebCore:

SVGFEImageElement should create the FEImage with either an Image or an
ImageBuffer based on whether the 'href' attribute points to a resource
image or an SVG element. For the SVG element, an ImageBuffer will be
created and the renderer of this element will be drawn to it.

Instead of holding a TreeScope and Image, FEImage can hold a variant of
Image and ImageBuffer.

Tests: svg/filters/feImage-element-primitive-subregion-expected.svg
       svg/filters/feImage-element-primitive-subregion.svg
       svg/filters/feImage-image-primitive-subregion-expected.svg
       svg/filters/feImage-image-primitive-subregion.svg

* svg/SVGFEImageElement.cpp:
(WebCore::SVGFEImageElement::imageBufferForEffect const):
(WebCore::SVGFEImageElement::build const):
* svg/SVGFEImageElement.h:
* svg/graphics/filters/SVGFEImage.cpp:
(WebCore::FEImage::create):
(WebCore::FEImage::FEImage):
(WebCore::FEImage::determineAbsolutePaintRect):
(WebCore::FEImage::platformApplySoftware):
(WebCore::FEImage::externalRepresentation const):
(WebCore::FEImage::createWithImage): Deleted.
(WebCore::FEImage::createWithIRIReference): Deleted.
(WebCore::FEImage::referencedRenderer const): Deleted.
* svg/graphics/filters/SVGFEImage.h:

LayoutTests:

* fast/shadow-dom/svg-feimage-href-in-shadow-tree.html:
The default primitive subregion of <feImage> is { -10%, -10%, 120%, 120% }
of the bounding box. So the filter result will be shifted (-10, -10) of
this test case if we do not specify the 'x' and then 'y' attributes.
Unlike the Image based feImage, the default behavior is to display the
ImageBuffer of the element with its logical size otherwise the drawing
will be pixelated if we try to fill the whole primitive subregion. To
make this test page matches its expected page, we need to specify the 'x'
and the 'y' attributes of the <feImage> to be zeros.

With and without specifying the 'x' and the 'y' attributes, the drawing
with this patch matches the drawing of this test page in other browsers.

* svg/filters/feImage-change-target-id-expected.txt:
* svg/filters/feImage-element-primitive-subregion-expected.svg: Added.
* svg/filters/feImage-element-primitive-subregion.svg: Added.
* svg/filters/feImage-image-primitive-subregion-expected.svg: Added.
* svg/filters/feImage-image-primitive-subregion.svg: Added.
* svg/filters/feImage-remove-target-expected.txt:



Canonical link: https://commits.webkit.org/244005@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285481 268f45cc-cd09-0410-ab3c-d52691b4dbfc