[css-grid] update the content-sized grid width before laying out a grid item with block constraints and aspect-ratio https://bugs.webkit.org/show_bug.cgi?id=231802
Reviewed by Javier Fernandez.
Source/WebCore:

For a grid item with an aspect-ratio, if it has block-constraints such as the relative logical height
case we consider in this CL, it should try and resolve it if possible and transfer this size into
the inline direction for the min/max content size. For the case that the grid width is content sized,
we need to update the width before laying out the grid items. Since the min-content contribution of
the grid item has changed based on the row sizes calculated in step 2 of sizing algorithm, we also
need to repeat the sizing algorithm steps to update the width of the track sizes.

* rendering/GridLayoutFunctions.cpp:
(WebCore::GridLayoutFunctions::isAspectRatioBlockSizeDependentChild):
* rendering/GridLayoutFunctions.h:
* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::layoutBlock):
(WebCore::RenderGrid::updateGridAreaForAspectRatioItems):
* rendering/RenderGrid.h:

LayoutTests:

Unskip the tests that are now passing and add a new test to check track size updates.

* imported/w3c/web-platform-tests/css/css-grid/grid-items/aspect-ratio-005-expected.html: Added.
* imported/w3c/web-platform-tests/css/css-grid/grid-items/aspect-ratio-005.html: Added.
* TestExpectations:
* platform/ios-wk2/TestExpectations:



Canonical link: https://commits.webkit.org/244021@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285497 268f45cc-cd09-0410-ab3c-d52691b4dbfc