Fix !ENABLE(ACCESSIBILITY) after r285427 https://bugs.webkit.org/show_bug.cgi?id=232893 <rdar://problem/85210424>
Reviewed by Andres Gonzalez.

In r285399 a number of AXObjectCache methods were affected by a removal of an
ENABLE(ACCESSIBILITY) guard. The definitions were pushed down to the large
!ENABLE(ACCESSIBILITY) block at the bottom of the file but `inline static` isn't valid
code so the compile issue was fixed in r285427. However the commits had also introduced a
linker error around AXObjectCache::accessibilityEnabled and
AXObjectCache::accessibilityEnhancedUserInterfaceEnabled.

To fix the static member variables referenced in those methods are `static constexpr` when
!ENABLE(ACCESSIBILITY).

* accessibility/AXObjectCache.h:



Canonical link: https://commits.webkit.org/244041@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285524 268f45cc-cd09-0410-ab3c-d52691b4dbfc