[WebAuthn] User handle is not saved on create via ASC https://bugs.webkit.org/show_bug.cgi?id=232900 rdar://85216105
Reviewed by Brent Fulgham.

This value is required to be stored along the credential on create calls
such that it can be returned via get calls. Currently, it is always read
as empty because user.id is empty after ipc, while the id is decoded to
idVector.

* UIProcess/WebAuthentication/Cocoa/WebAuthenticatorCoordinatorProxy.mm:
(WebKit::configureRegistrationRequestContext):



Canonical link: https://commits.webkit.org/244054@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285537 268f45cc-cd09-0410-ab3c-d52691b4dbfc