imported/w3c/web-platform-tests/webmessaging/broadcastchannel/workers.html is flaky crashing in debug https://bugs.webkit.org/show_bug.cgi?id=232920
Reviewed by Alex Christensen.

When WorkerGlobalScope::postTask() gets called, the task may get destroyed on the worker thread, without
getting executed in the case where the worker thread is about to exit. This was causing trouble in
BroadcastChannel::dispatchMessageTo() where we were calling WorkerGlobalScope::postTask() and capturing
a CallbackAggregator. We were relying on the task actually executing to dispatch the CallbackAggregator
back to the maint thread so that the completion handler is always called on the main thread.

To address the issue, we now capture a WTF::ScopeExit which calls the completion handler on the main
thread upon destruction. This way, the completion handler will always get called on the main thread,
no matter what.

* dom/BroadcastChannel.cpp:
(WebCore::BroadcastChannel::dispatchMessageTo):



Canonical link: https://commits.webkit.org/244095@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285587 268f45cc-cd09-0410-ab3c-d52691b4dbfc