[LFC][IFC] Add support for non-initial text-alignment values on bidi content https://bugs.webkit.org/show_bug.cgi?id=232968
Reviewed by Antti Koivisto.

The text-align property "moves" the root inline box as a whole and not the individual runs.
This patch ensures that this offset value is taken into account when computing the horizontal position for the visually ordered runs too.
(Note that this offset value is baked into the logical position of the runs, see LineBox::logicalRectForTextRun.)

* layout/formattingContexts/inline/InlineDisplayContentBuilder.cpp:
(WebCore::Layout::InlineDisplayContentBuilder::createBoxesAndUpdateGeometryForLineContent):



Canonical link: https://commits.webkit.org/244129@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285628 268f45cc-cd09-0410-ab3c-d52691b4dbfc