-Warray-bounds, -Wstringop-truncation, -Wstringop-overread warnings in Packed.h https://bugs.webkit.org/show_bug.cgi?id=226557 <rdar://problem/79103658>
Patch by Michael Catanzaro <mcatanzaro@gnome.org> on 2021-11-11
Reviewed by Darin Adler.

.:

Suppress -Wstringop-overread for all of WebKit when building with GCC 11 or newer. This is
an unfortunate way to deal with false positives that we ought to be able to silence in just
a few places, but when warnings do not respect pragmas we have few alternatives, and trying
to remove derived sources from the unified build in order to build them with special flags
is just not worth the effort.

This is part of a larger trend of oversensitive warnings from GCC 11.

* Source/cmake/WebKitCompilerFlags.cmake:

Source/WebCore:

Since -Wstringop-overread is now suppressed globally, we can include WebSocket.cpp in the
unified build once again.

* CMakeLists.txt:
* Sources.txt:
* WebCore.xcodeproj/project.pbxproj:


Canonical link: https://commits.webkit.org/244150@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285652 268f45cc-cd09-0410-ab3c-d52691b4dbfc