[JSC][32bit] Unskip JSTests/microbenchmarks/redefine-property-accessor-dictionary.js https://bugs.webkit.org/show_bug.cgi?id=233034
Unreviewed gardening.

This is skipped for memoryLimited *and* mips|arm systems, but a)
seems to run fine on both setups and b) is almost identical to the
other tests/benchmarks that landed in the same patch, that do not
have that limitation. Optimistically remove the skip and let's see
if it sticks.

Patch by Xan Lopez <xan@igalia.com> on 2021-11-12

* microbenchmarks/redefine-property-accessor-dictionary.js:


Canonical link: https://commits.webkit.org/244171@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285712 268f45cc-cd09-0410-ab3c-d52691b4dbfc