REGRESSION(r285481): Infinite recursion with cyclic filter reference https://bugs.webkit.org/show_bug.cgi?id=232972 rdar://85264240
Reviewed by Wenson Hsieh.

Source/WebCore:

Before r285481, we were creating the ImageBuffer of the referenced SVGElement
for the FEImage through RenderSVGResourceFilter::postApplyResource(). Now
we create this ImageBuffer through RenderSVGResourceFilter::applyResource().
The difference is at the end of RenderSVGResourceFilter::applyResource()
we add an entry to m_rendererFilterDataMap. This entry was preventing
trying to rebuild the SVGFilter for the same renderer if there is a
cyclic reference.

The fix is to add the entry in m_rendererFilterDataMap before creating the
SVGFilter. If an error happens, this entry will be removed before returning.

Test: svg/filters/feImage-cyclic-reference.svg

* rendering/svg/RenderSVGResourceFilter.cpp:
(WebCore::RenderSVGResourceFilter::applyResource):

LayoutTests:

* svg/filters/feImage-cyclic-reference-expected.txt: Added.
* svg/filters/feImage-cyclic-reference.svg: Added.



Canonical link: https://commits.webkit.org/244213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285769 268f45cc-cd09-0410-ab3c-d52691b4dbfc