Make sort-Xcode-project-file idempotent https://bugs.webkit.org/show_bug.cgi?id=233077
Reviewed by Wenson Hsieh.

Source/bmalloc:

* libpas/libpas.xcodeproj/project.pbxproj: Sorted.

Source/JavaScriptCore:

* JavaScriptCore.xcodeproj/project.pbxproj: Sorted.

Source/WebCore:

* WebCore.xcodeproj/project.pbxproj: Sorted.

Source/WebKit:

* WebKit.xcodeproj/project.pbxproj: Sorted.

Source/WebKitLegacy:

* WebKitLegacy.xcodeproj/project.pbxproj: Sorted.

Tools:

* Scripts/sort-Xcode-project-file:
(sortChildrenByFileName): Fix the "directories first" rule to work in a straightforward
way that is symmetric, and make the rest of the rules fall through when two files are
equal for the same reason. Also remove almost all names from the special list of filenames
to treat as files even though they have no extension. For example, "jsc" is a directory
in WebCore and it's not needed to sort products correctly in JavaScriptCore. And
"minidom" is both a product and a directory.

* DumpRenderTree/DumpRenderTree.xcodeproj/project.pbxproj: Sorted.
* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj: Ditto.
* WebKitTestRunner/WebKitTestRunner.xcodeproj/project.pbxproj: Ditto.


Canonical link: https://commits.webkit.org/244214@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285770 268f45cc-cd09-0410-ab3c-d52691b4dbfc