[CSS contain] The size should be updated if layout containment is inside a fit-content parent https://bugs.webkit.org/show_bug.cgi?id=233046
Reviewed by Rob Buis.

Source/WebCore:

When the object is a layout containment, we also need to add size containment as the constraint of
being a relayout boundary. Otherwise when the layout size depends on the content, the layout size
might not be updated correctly.

Test: imported/w3c/web-platform-tests/css/css-contain/contain-layout-019.html

* rendering/RenderObject.cpp:
(WebCore::objectIsRelayoutBoundary):

LayoutTests:

* TestExpectations:



Canonical link: https://commits.webkit.org/244280@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285854 268f45cc-cd09-0410-ab3c-d52691b4dbfc