Fix for accessibility/notification-listeners.html in isolated tree mode. https://bugs.webkit.org/show_bug.cgi?id=233228 <rdar://problem/85484341>
Reviewed by Chris Fleizach.

Made this test async in order to pass in isolated tree mode.
The first accessible element needs to be retrieved with waitForElementById,
otherwise if we use accessibilityController.accessibleElementById, the
test would fail randomly (~10 out of 1000 flakiness). It requires
further investigation to determine if this is an issue with the <select>
element or with the first accessible element the test retrieves.

* accessibility/notification-listeners.html:
* platform/mac/accessibility/notification-listeners-expected.txt:



Canonical link: https://commits.webkit.org/244333@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285924 268f45cc-cd09-0410-ab3c-d52691b4dbfc