Audio Rate Gets Messed up With Safari WebRTC and Bluetooth Switching https://bugs.webkit.org/show_bug.cgi?id=232822 <rdar://problem/85418545>
Reviewed by Eric Carlson.

We usually create the unit, start the unit and stop the unit.
After that cycle, if we restart the unit, we expect the configuration to stay the same in WebProcess.
But, in LocalAudioMediaStreamTrackRendererInternalUnit::stop, we were disposing the audio unit and recreating it if necesary in LocalAudioMediaStreamTrackRendererInternalUnit::start.
If the sample rate changed, the new audio unit would use the new sample rate while WebProcess will use the old sample rate.
To prevent this, we now always reuse the same description to initialize the audio unit.

Manually tested.

* platform/mediastream/cocoa/AudioMediaStreamTrackRendererInternalUnit.cpp:
(WebCore::LocalAudioMediaStreamTrackRendererInternalUnit::createAudioUnitIfNeeded):



Canonical link: https://commits.webkit.org/244382@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@285985 268f45cc-cd09-0410-ab3c-d52691b4dbfc