[Flatpak SDK] Add build recipe for libjxl https://bugs.webkit.org/show_bug.cgi?id=233326
Reviewed by Philippe Normand.

Add the build recipe for libjxl, which in turns needs highway (libhwy), which gets imported
as well. Highway is always built as a static library, and therefore only needed at build
time. As for libjxl, the recipe tries to disable as much as possible of bundled sources from
its third_party/ subdirectory, but LodePNG gets always needed, and the same goes for either
SkCMS or LCMS2; and in both cases they get built and linked statically into the resulting
libjxl library. Ideally we would prefer that libjxl could use a system-installed copy of
LCMS2 because that is already available in the SDK, but that not being an option SkCMS is
preferred as it is smaller.

* elements/sdk-platform.bst: List libjxl for installation in the SDK.
* elements/sdk/highway.bst: Added.
* elements/sdk/libjxl.bst: Added.



Canonical link: https://commits.webkit.org/244444@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286057 268f45cc-cd09-0410-ab3c-d52691b4dbfc