[Flatpak SDK] Add build recipe for libjxl, fix after r286057 https://bugs.webkit.org/show_bug.cgi?id=233326
Reviewed by Philippe Normand.

Fix the generated .pc files installed by libjxl, which wrongly include libhwy as a
requirement, even when it is used internally in the implementation only and linked
in statically. This solves a remaining issue that CMake would not be able to find
the library using pkg-config.

* elements/sdk/libjxl.bst: Edit installed .pc files to remove libhwy.



Canonical link: https://commits.webkit.org/244514@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286127 268f45cc-cd09-0410-ab3c-d52691b4dbfc