WKContentView should allow -captureTextFromCamera: with a caret selection unless the sender is the callout bar https://bugs.webkit.org/show_bug.cgi?id=233378 rdar://85478344
Reviewed by Tim Horton.

Source/WebKit:

Allow clients to query whether or not the "capture text from camera" action should be enabled on WKWebView in
contexts outside of the callout bar. Currently, WKContentView returns NO for this action when there is a non-
collapsed selection range, but this behavior is only intended for the callout bar; in the case where the action
is (for instance) triggered by a button in the web view chrome, it doesn't make sense to only enable it when the
selection is a caret.

Additionally, adjust the check to use focused element information instead of editor state, such that WebKit
clients can easily query this API (`-canPerformAction:withSender:`) inside of input delegate callbacks when
beginning an input session to know whether or not "capture text from camera" should be allowed.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView canPerformActionForWebView:withSender:]):

Tools:

Adjust an API test to exercise the changes. See WebKit/ChangeLog for more details.

* TestWebKitAPI/Tests/WebKitCocoa/WKContentViewEditingActions.mm:



Canonical link: https://commits.webkit.org/244623@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286260 268f45cc-cd09-0410-ab3c-d52691b4dbfc