WebSWServerConnection::createFetchTask does not need to query twice for a registration https://bugs.webkit.org/show_bug.cgi?id=233467
Reviewed by Chris Dumez.

Instead of searching for a worker and then searching for a registration,
we now search for a registration and, if not null, directly get its active worker.
This makes it clear the registration is not null in follow-up code and removes a redundant query.

No change of behavior, this is a refactoring.

* NetworkProcess/ServiceWorker/WebSWServerConnection.cpp:
(WebKit::WebSWServerConnection::createFetchTask):



Canonical link: https://commits.webkit.org/244718@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@286359 268f45cc-cd09-0410-ab3c-d52691b4dbfc